+++ BackEnd/b2formeditor/src/main/java/com/b2formeditor/services/CrudService.java
@@ -9,6 +9,6 @@ import java.util.List;
 public interface CrudService<T> {
     T save(T entity);
     List<T> getAll();
-    T getById(Long id);
-    void delete(Long id);
+    T getById(Integer id);
+    void delete(Integer id);
 }
